client_pc/requirements.txt
# Python dependencies for Windows PC RTMP client

# YAML configuration parsing
PyYAML>=6.0

# System utilities (disk space monitoring)
psutil>=5.9.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# GUI framework
PyQt5>=5.15.0

# Numerical computing (required by opencv)
# Python 3.14 requires numpy 2.x
numpy>=2.0.0,<2.3.0  # Compatible with opencv-python

# Video processing
opencv-python>=4.12.0

# Image processing
Pillow==11.3.0

# Global hotkeys
pynput==1.8.1
